Add generic interface for count distinct
